+++ tests/configs/h2o_configs/test_custom.py
@@ -7,7 +7,7 @@ import pydantic as pyd
 import pathlib
 import pytest
 
-from student_success_tool.configs.h2o import custom
+from student_success_tool.configs.h2o_configs import custom
 
 
 SRC_ROOT = (
